reverted back to having intern_id

+++ codeartproject/projectFiles/endpoints/internships.py
@@ -26,6 +26,7 @@ class InternModify(Resource):
     parser.add_argument('start_datetime', type=str)
     parser.add_argument('end_datetime', type=str)
     parser.add_argument('details', type=str)
+    parser.add_argument('intern_id', type=str)
 
     def put(self):
 
@@ -33,7 +34,7 @@ class InternModify(Resource):
             args = self.parser.parse_args()
             if verify_auth(args['auth'], args['id']):
 
-                internship = Internship.query.get(args["id"])
+                internship = Internship.query.get(args["intern_id"])
                 
                 if internship:
 
@@ -70,25 +71,24 @@ class InternCreate(Resource):
     parser.add_argument('details', type=str)
     
     def post(self):
+        print(request.data)
 
         try:
             created_id = uuid.uuid4()
             args = self.parser.parse_args()
-            print("Before Verified")
             if verify_auth('auth', 'id'):
-                print("Verifed")
                 intern_id = str( created_id )
-                #acc = Account.query.get( str( args["id"] ) )
-                #if acc:
-                internship = Internship(id=intern_id, location=args["location"], company=args["company"], role=args["role"], 
-                link=args["link"], start_datetime=args["start_datetime"], end_datetime=args["end_datetime"], details=args["details"])
+                acc = Account.query.get( str( args["id"] ) )
+                if acc:
+                    internship = Internship(id=intern_id, location=args["location"], company=args["company"], role=args["role"], 
+                    link=args["link"], start_datetime=args["start_datetime"], end_datetime=args["end_datetime"], details=args["details"])
 
-                db.session.add(internship)
-                db.session.commit()
+                    db.session.add(internship)
+                    db.session.commit()
 
-                return {"success": True}, 201
-                #else:
-                    #return {"msg":"No Account","success":False}, 400
+                    return {"success": True}, 201
+                else:
+                    return {"msg":"No Account","success":False}, 400
             else:
                 return {"msg":"Invalid ID or Auth Token","success":False}, 400
         except Exception as exe:
@@ -101,10 +101,11 @@ class InternInfo(Resource):
             parser = reqparse.RequestParser()
             parser.add_argument('id', type=str)
             parser.add_argument('auth', type=str)
+            parser.add_argument('intern_id', type=str)
             args = parser.parse_args()
 
             if verify_auth('auth', 'id'):
-                internship = Internship.query.get( args["id"] )
+                internship = Internship.query.get( args["intern_id"] )
                 if internship:
                     print("Internship Exists")
 
@@ -123,20 +124,20 @@ class InternDelete(Resource):
         parser = reqparse.RequestParser()
         parser.add_argument('id', type=str)
         parser.add_argument('auth', type=str)
+        parser.add_argument('intern_id', type=str)
 
         try:
             args = parser.parse_args()
-            #if verify_auth('auth', 'id'):
-                #acc = Account.query.get( str( args["id"] ) )
-                #if acc:
-            internship = Internship.query.get(args["id"])
-
-            if internship:
-                db.session.delete(internship)
-                db.session.commit()
-                return {"msg":"Internship Deleted","success":True}, 200
-                #else:
-                    #return {"msg":"No Account","success":False}, 400
+            if verify_auth('auth', 'id'):
+                acc = Account.query.get( str( args["id"] ) )
+                if acc:
+                    internship = Internship.query.get(args["intern_id"])
+                    if internship:
+                        db.session.delete(internship)
+                        db.session.commit()
+                        return {"msg":"Internship Deleted","success":True}, 200
+                else:
+                    return {"msg":"No Account","success":False}, 400
             else:
                 return {"msg":"Internship Could Not Be Deleted","success":False}, 400
         except Exception as exe:
